requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.4
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==3.1.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.2
packaging==20.7
pandas==0.24.1
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.0
python-dotenv==0.13.0
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.2
toml==0.10.2
urllib3==1.24.1
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.3.1
zipp==3.4.0
